Pallette: add a version field
to track palette changes requiring a texture cache flush
beware that the `col` field is still public (FIXME)

fixes various palette issues in SDL 2
issue #113

+++ gemrb/plugins/SDLVideo/SDL12Video.cpp
@@ -318,7 +318,7 @@ void SDL12VideoDriver::BlitSpriteNativeClipped(const Sprite2D* spr, const Sprite
 	if (surf->format->BytesPerPixel == 1) {
 		c.a = 255; // FIXME: this is probably actually contigent on something else...
 
-		const unsigned int shaderflags = (BLIT_TINTED|BLIT_GREY|BLIT_SEPIA);
+		const short int shaderflags = (BLIT_TINTED|BLIT_GREY|BLIT_SEPIA);
 		unsigned int version = remflags&shaderflags;
 		RenderSpriteVersion(sdlspr, version, &c);
 		// since the "shading" has been done we clear the flags
